refactor: move filters to new file
